+++ content/00-namespace.js
@@ -56,6 +56,7 @@
         keepSessionAlive: true,
         keepScreenAwake: true,
         showProfessorRatings: true,
+        showSmartSchedulePlanner: true,
       },
       selectedPassId: null,
       cachedRegisterButton: null,